Explicitly cast values to integers
These methods should return integers according to their type information,
however in practice we are sometimes getting string values come back. The
proper fix for this would be to move the type information from the docblock to
PHP native types, which would enforce the declared return type. Doing that is
out of scope for this task.